Week3 (11-15)/Day 12/react_fastapi_challenge/fastapi_company_app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
pydantic
